refactor(server): extract AI scan pipeline to internal/aiscan/
Move ai_scan_pipeline.go to internal/aiscan/pipeline.go. Remove the unused
*Server field from PipelineManager (pm.server was stored but never read).
Rename aiScanPipelineStore to aiscan.Store (exported package interface).
Update call site in internal/server/ to drop the server argument.

Also move cross_validation.go to internal/aiscan/ (no HTTP deps, called
directly by PipelineManager). Move associated test files to internal/aiscan/.

Refs: PKG-2
